refactor: move auth session admin helpers
