+++ src/stdlib/auth.rs
@@ -67,7 +67,10 @@ use sessions::{
     build_rotated_session, get_session_by_id, migrate_session, update_session_data,
     update_session_tokens,
 };
-pub use sessions::{cleanup_expired_sessions, delete_session_by_id, store_session};
+pub use sessions::{
+    cleanup_expired_sessions, delete_all_sessions_for_user, delete_session_by_id,
+    get_sessions_for_user, store_session,
+};
 #[cfg(test)]
 use storage::store_auth_challenge_sqlite;
 use storage::{
@@ -2211,324 +2214,6 @@ fn cleanup_expired_exchange_tokens_postgres(cutoff: i64) -> std::result::Result<
             &[&cutoff],
         )
         .map_err(|e| e.to_string())?;
-
-    Ok(count)
-}
-
-/// Get all sessions for a user
-pub fn get_sessions_for_user(
-    user_id: &str,
-    current_session_id: Option<&str>,
-) -> std::result::Result<Vec<SessionInfo>, String> {
-    let now = chrono::Utc::now().timestamp();
-    let config = get_auth_config();
-    let store_type = config.as_ref().map(|c| &c.session_store);
-
-    match store_type {
-        Some(SessionStore::Sqlite(_)) => {
-            get_sessions_for_user_sqlite(user_id, current_session_id, now)
-        }
-        Some(SessionStore::Postgres(_)) => {
-            get_sessions_for_user_postgres(user_id, current_session_id, now)
-        }
-        Some(SessionStore::Redis(_)) => {
-            get_sessions_for_user_redis(user_id, current_session_id, now)
-        }
-        _ => {
-            // Memory backend
-            let store = SESSION_STORE.lock().unwrap();
-            Ok(store.get_sessions_for_user(user_id, current_session_id, now))
-        }
-    }
-}
-
-fn get_sessions_for_user_sqlite(
-    user_id: &str,
-    current_session_id: Option<&str>,
-    now: i64,
-) -> std::result::Result<Vec<SessionInfo>, String> {
-    let conn_guard = SQLITE_CONN.lock().unwrap();
-    let conn = conn_guard.as_ref().ok_or("SQLite not initialized")?;
-
-    let mut stmt = conn
-        .prepare(
-            "SELECT id, user_id, provider, created_at, expires_at FROM auth_sessions
-         WHERE user_id = ?1 AND expires_at > ?2 ORDER BY created_at DESC",
-        )
-        .map_err(|e| e.to_string())?;
-
-    let rows = stmt
-        .query_map(rusqlite::params![user_id, now], |row| {
-            Ok(SessionInfo {
-                id: row.get(0)?,
-                user_id: row.get(1)?,
-                provider: row.get(2)?,
-                created_at: row.get(3)?,
-                expires_at: row.get(4)?,
-                is_current: false,
-            })
-        })
-        .map_err(|e| e.to_string())?;
-
-    let mut sessions: Vec<SessionInfo> = rows.filter_map(|r| r.ok()).collect();
-
-    // Mark current session
-    if let Some(current_id) = current_session_id {
-        for session in &mut sessions {
-            if session.id == current_id {
-                session.is_current = true;
-            }
-        }
-    }
-
-    Ok(sessions)
-}
-
-fn get_sessions_for_user_postgres(
-    user_id: &str,
-    current_session_id: Option<&str>,
-    now: i64,
-) -> std::result::Result<Vec<SessionInfo>, String> {
-    let url_guard = POSTGRES_URL.lock().unwrap();
-    let url = url_guard.as_ref().ok_or("PostgreSQL not initialized")?;
-
-    let mut client = postgres::Client::connect(url, postgres::NoTls).map_err(|e| e.to_string())?;
-
-    let rows = client
-        .query(
-            "SELECT id, user_id, provider, created_at, expires_at FROM auth_sessions
-         WHERE user_id = $1 AND expires_at > $2 ORDER BY created_at DESC",
-            &[&user_id, &now],
-        )
-        .map_err(|e| e.to_string())?;
-
-    let mut sessions: Vec<SessionInfo> = rows
-        .iter()
-        .map(|row| SessionInfo {
-            id: row.get(0),
-            user_id: row.get(1),
-            provider: row.get(2),
-            created_at: row.get(3),
-            expires_at: row.get(4),
-            is_current: current_session_id
-                .map(|c| c == row.get::<_, String>(0))
-                .unwrap_or(false),
-        })
-        .collect();
-
-    // Mark current session
-    if let Some(current_id) = current_session_id {
-        for session in &mut sessions {
-            if session.id == current_id {
-                session.is_current = true;
-            }
-        }
-    }
-
-    Ok(sessions)
-}
-
-fn get_sessions_for_user_redis(
-    user_id: &str,
-    current_session_id: Option<&str>,
-    now: i64,
-) -> std::result::Result<Vec<SessionInfo>, String> {
-    let url_guard = REDIS_URL.lock().unwrap();
-    let url = url_guard.as_ref().ok_or("Redis not initialized")?;
-
-    let client =
-        redis::Client::open(url.as_str()).map_err(|e| format!("Redis client error: {}", e))?;
-    let mut conn = client
-        .get_connection()
-        .map_err(|e| format!("Redis connection error: {}", e))?;
-
-    let mut sessions = Vec::new();
-    let mut cursor = 0u64;
-
-    loop {
-        let (new_cursor, keys): (u64, Vec<String>) = redis::cmd("SCAN")
-            .arg(cursor)
-            .arg("MATCH")
-            .arg("ntnt:session:*")
-            .arg("COUNT")
-            .arg(100)
-            .query(&mut conn)
-            .map_err(|e| format!("Redis SCAN error: {}", e))?;
-
-        for key in keys {
-            let result: Option<String> = redis::cmd("GET")
-                .arg(&key)
-                .query(&mut conn)
-                .map_err(|e| format!("Redis GET error: {}", e))?;
-
-            if let Some(json_str) = result {
-                if let Ok(json) = serde_json::from_str::<serde_json::Value>(&json_str) {
-                    let session_user_id = json["user_id"].as_str().unwrap_or("");
-                    let expires_at = json["expires_at"].as_i64().unwrap_or(0);
-
-                    if session_user_id == user_id && expires_at > now {
-                        let session_id = json["id"].as_str().unwrap_or("").to_string();
-                        sessions.push(SessionInfo {
-                            id: session_id.clone(),
-                            user_id: session_user_id.to_string(),
-                            provider: json["provider"].as_str().unwrap_or("").to_string(),
-                            created_at: json["created_at"].as_i64().unwrap_or(0),
-                            expires_at,
-                            is_current: current_session_id
-                                .map(|c| c == session_id)
-                                .unwrap_or(false),
-                        });
-                    }
-                }
-            }
-        }
-
-        cursor = new_cursor;
-        if cursor == 0 {
-            break;
-        }
-    }
-
-    // Sort by created_at descending
-    sessions.sort_by(|a, b| b.created_at.cmp(&a.created_at));
-
-    Ok(sessions)
-}
-
-/// Delete all sessions for a user, optionally keeping one session
-/// Returns the number of sessions deleted
-pub fn delete_all_sessions_for_user(
-    user_id: &str,
-    keep_session_id: Option<&str>,
-) -> std::result::Result<u64, String> {
-    let config = get_auth_config();
-    let store_type = config.as_ref().map(|c| &c.session_store);
-
-    match store_type {
-        Some(SessionStore::Sqlite(_)) => {
-            delete_all_sessions_for_user_sqlite(user_id, keep_session_id)
-        }
-        Some(SessionStore::Postgres(_)) => {
-            delete_all_sessions_for_user_postgres(user_id, keep_session_id)
-        }
-        Some(SessionStore::Redis(_)) => {
-            delete_all_sessions_for_user_redis(user_id, keep_session_id)
-        }
-        _ => {
-            // Memory backend
-            let mut store = SESSION_STORE.lock().unwrap();
-            Ok(store.delete_all_sessions_for_user(user_id, keep_session_id) as u64)
-        }
-    }
-}
-
-fn delete_all_sessions_for_user_sqlite(
-    user_id: &str,
-    keep_session_id: Option<&str>,
-) -> std::result::Result<u64, String> {
-    let conn_guard = SQLITE_CONN.lock().unwrap();
-    let conn = conn_guard.as_ref().ok_or("SQLite not initialized")?;
-
-    let count = if let Some(keep_id) = keep_session_id {
-        conn.execute(
-            "DELETE FROM auth_sessions WHERE user_id = ?1 AND id != ?2",
-            rusqlite::params![user_id, keep_id],
-        )
-        .map_err(|e| e.to_string())?
-    } else {
-        conn.execute(
-            "DELETE FROM auth_sessions WHERE user_id = ?1",
-            rusqlite::params![user_id],
-        )
-        .map_err(|e| e.to_string())?
-    };
-
-    Ok(count as u64)
-}
-
-fn delete_all_sessions_for_user_postgres(
-    user_id: &str,
-    keep_session_id: Option<&str>,
-) -> std::result::Result<u64, String> {
-    let url_guard = POSTGRES_URL.lock().unwrap();
-    let url = url_guard.as_ref().ok_or("PostgreSQL not initialized")?;
-
-    let mut client = postgres::Client::connect(url, postgres::NoTls).map_err(|e| e.to_string())?;
-
-    let count = if let Some(keep_id) = keep_session_id {
-        client
-            .execute(
-                "DELETE FROM auth_sessions WHERE user_id = $1 AND id != $2",
-                &[&user_id, &keep_id],
-            )
-            .map_err(|e| e.to_string())?
-    } else {
-        client
-            .execute("DELETE FROM auth_sessions WHERE user_id = $1", &[&user_id])
-            .map_err(|e| e.to_string())?
-    };
-
-    Ok(count)
-}
-
-fn delete_all_sessions_for_user_redis(
-    user_id: &str,
-    keep_session_id: Option<&str>,
-) -> std::result::Result<u64, String> {
-    let url_guard = REDIS_URL.lock().unwrap();
-    let url = url_guard.as_ref().ok_or("Redis not initialized")?;
-
-    let client =
-        redis::Client::open(url.as_str()).map_err(|e| format!("Redis client error: {}", e))?;
-    let mut conn = client
-        .get_connection()
-        .map_err(|e| format!("Redis connection error: {}", e))?;
-
-    let mut count = 0u64;
-    let mut cursor = 0u64;
-
-    loop {
-        let (new_cursor, keys): (u64, Vec<String>) = redis::cmd("SCAN")
-            .arg(cursor)
-            .arg("MATCH")
-            .arg("ntnt:session:*")
-            .arg("COUNT")
-            .arg(100)
-            .query(&mut conn)
-            .map_err(|e| format!("Redis SCAN error: {}", e))?;
-
-        for key in keys {
-            let result: Option<String> = redis::cmd("GET")
-                .arg(&key)
-                .query(&mut conn)
-                .map_err(|e| format!("Redis GET error: {}", e))?;
-
-            if let Some(json_str) = result {
-                if let Ok(json) = serde_json::from_str::<serde_json::Value>(&json_str) {
-                    let session_user_id = json["user_id"].as_str().unwrap_or("");
-                    let session_id = json["id"].as_str().unwrap_or("");
-
-                    if session_user_id == user_id {
-                        // Skip the session we want to keep
-                        if keep_session_id.map(|k| k == session_id).unwrap_or(false) {
-                            continue;
-                        }
-
-                        let _: () = redis::cmd("DEL")
-                            .arg(&key)
-                            .query(&mut conn)
-                            .map_err(|e| format!("Redis DEL error: {}", e))?;
-                        count += 1;
-                    }
-                }
-            }
-        }
-
-        cursor = new_cursor;
-        if cursor == 0 {
-            break;
-        }
-    }
 
     Ok(count)
 }
